move config to src and types in definitation
