+++ .gitignore
@@ -1,3 +1,3 @@
-config.js
+src/config.js
 node_modules/
 dist/
